requirements.txt
aiofile==3.9.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
Authlib==1.6.10
beartype==0.22.9
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
cssselect==1.4.0
cssutils==2.14.0
cyclopts==4.10.2
dnspython==2.8.0
docstring_parser==0.18.0
docutils==0.22.4
email-validator==2.3.0
encutils==1.0.0
exceptiongroup==1.3.1
fastmcp==3.2.4
greenlet==3.4.0
griffelib==2.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
lxml==6.1.0
markdown-it-py==4.0.0
mcp==1.27.0
mdurl==0.1.2
more-itertools==11.0.2
msvc_runtime==14.44.35112
openapi-pydantic==0.5.1
opentelemetry-api==1.41.0
packaging==26.1
pathable==0.5.0
platformdirs==4.9.6
playwright==1.58.0
premailer==3.10.0
py-key-value-aio==0.4.4
pycparser==3.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
pyee==13.0.1
Pygments==2.20.0
PyJWT==2.12.1
pyperclip==1.11.0
python-dotenv==1.2.2
python-multipart==0.0.26
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rich==15.0.0
rich-rst==1.3.2
rpds-py==0.30.0
sse-starlette==3.3.4
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.3.1
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
yagmail==0.15.293
zipp==3.23.1
